# docs/talent-x-openapi.yaml
...
openapi: 3.1.0
info:
  title: Talent-X API
  version: 1.0.0
  description: 'Contrat d''API REST de Talent-X — source de vérité dérivée de TX-SPEC-002 v2.0 (catalogue
    d''endpoints, matrice d''autorisation, conventions et format d''erreur) et alignée sur le modèle de
    données TX-DATA-006.


    Conventions transverses : préfixe /api/v1 ; jeton d''accès JWT (RS256) via en-tête Authorization ;
    pagination par enveloppe { data, meta } ; idempotence des écritures sensibles via Idempotency-Key
    ; opérations longues asynchrones (202 + ressource de statut) ; rate limiting signalé par les en-têtes
    RateLimit-*. L''autorisation combine rôle, appartenance (lien coach↔athlète), propriété et consentement
    ; voir TX-SPEC-002 §6.'
  contact:
    name: Talent-X
  license:
    name: Propriétaire
servers:
- url: https://api.talent-x.example/api/v1
  description: Production (URL à confirmer)
- url: http://localhost:3000/api/v1
  description: Développement local
...
security:
- bearerAuth: []
paths:
...
  /competitions:
    post:
      tags:
      - Compétitions
      summary: Créer une compétition
      operationId: createCompetition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitionCreate'
      responses:
        '201':
          description: Compétition créée.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competition'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
        '403':
          $ref: '#/components/responses/Forbidden'
    get:
      tags:
      - Compétitions
      summary: 'Lister les compétitions (coach : les siennes ; athlète : engagées)'
      operationId: listCompetitions
      parameters:
      - $ref: '#/components/parameters/PageParam'
      - $ref: '#/components/parameters/LimitParam'
      - $ref: '#/components/parameters/SortParam'
      - $ref: '#/components/parameters/CompetitionStatusFilter'
      responses:
        '200':
          description: Liste paginée des compétitions.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionPage'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
  /competitions/{id}:
    get:
      tags:
      - Compétitions
      summary: Lire une compétition autorisée
      operationId: getCompetition
      parameters:
      - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Compétition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competition'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    put:
      tags:
      - Compétitions
      summary: Modifier une compétition
      operationId: updateCompetition
      parameters:
      - $ref: '#/components/parameters/IdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitionUpdate'
      responses:
        '200':
          description: Compétition mise à jour.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competition'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
      - Compétitions
      summary: Supprimer (logique) une compétition
      operationId: deleteCompetition
      parameters:
      - $ref: '#/components/parameters/IdPath'
      responses:
        '204':
          description: Compétition supprimée.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /competitions/{id}/entries:
    post:
      tags:
      - Compétitions
      summary: Engager des athlètes à une compétition
      operationId: engageAthletes
      parameters:
      - $ref: '#/components/parameters/IdPath'
      - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngageRequest'
      responses:
        '201':
          description: Engagements créés.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionEntryList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
    get:
      tags:
      - Compétitions
      summary: Lister les engagements d'une compétition
      operationId: listEntries
      parameters:
      - $ref: '#/components/parameters/IdPath'
      responses:
        '200':
          description: Engagements de la compétition.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionEntryList'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /competitions/{id}/entries/{entryId}:
    delete:
      tags:
      - Compétitions
      summary: Désengager un athlète
      operationId: unengageAthlete
      parameters:
      - $ref: '#/components/parameters/IdPath'
      - name: entryId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '204':
          description: Engagement supprimé.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationFailed'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/ServerError'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token JWT signé RS256.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortParam:
      name: sort
      in: query
      required: false
      description: Champ de tri, préfixé de '-' pour l'ordre décroissant (ex. -createdAt).
      schema:
        type: string
...
    CompetitionStatusFilter:
      name: status
      in: query
      required: false
      schema:
        $ref: '#/components/schemas/CompetitionStatus'
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: Clé d'idempotence générée par le client pour absorber les reprises réseau.
      schema:
        type: string
  headers:
    RateLimit-Limit:
      description: Quota sur la fenêtre courante.
      schema:
        type: integer
    RateLimit-Remaining:
      description: Requêtes restantes sur la fenêtre.
      schema:
        type: integer
    RateLimit-Reset:
      description: Secondes avant réinitialisation du quota.
      schema:
        type: integer
    Retry-After:
      description: Secondes à attendre avant un nouvel essai.
      schema:
        type: integer
  responses:
    BadRequest:
      description: Requête invalide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 400
            error: BAD_REQUEST
            message: Invalid request
    Unauthorized:
      description: Authentification requise ou invalide.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 401
            error: UNAUTHORIZED
            message: Authentication required
    Forbidden:
      description: Accès refusé (rôle, appartenance ou propriété).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 403
            error: FORBIDDEN
            message: Access denied
    NotFound:
      description: Ressource introuvable.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 404
            error: NOT_FOUND
            message: Resource not found
...
    ValidationFailed:
      description: Échec de validation des données.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 422
            error: VALIDATION_ERROR
            message: Validation failed
            details:
            - field: email
              constraint: isEmail
              message: email must be a valid email address
    TooManyRequests:
      description: Trop de requêtes (rate limiting).
      headers:
        RateLimit-Limit:
          $ref: '#/components/headers/RateLimit-Limit'
        RateLimit-Remaining:
          $ref: '#/components/headers/RateLimit-Remaining'
        RateLimit-Reset:
          $ref: '#/components/headers/RateLimit-Reset'
        Retry-After:
          $ref: '#/components/headers/Retry-After'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 429
            error: RATE_LIMITED
            message: Too many requests
    ServerError:
      description: Erreur interne.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 500
            error: INTERNAL_ERROR
            message: Internal server error
...
    IdempotencyConflict:
      description: 'Conflit d''idempotence : la clé a déjà été utilisée avec une charge différente.'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            statusCode: 409
            error: IDEMPOTENCY_CONFLICT
            message: Idempotency key reused with a different payload
...
  schemas:
    Error:
      type: object
      properties:
        statusCode:
          type: integer
        error:
          type: string
          description: Code d'erreur stable et exploitable côté client.
        message:
          type: string
        details:
          type: array
          items:
            $ref: '#/components/schemas/ValidationDetail'
        timestamp:
          type: string
          format: date-time
        path:
          type: string
      required:
      - statusCode
      - error
      - message
    ValidationDetail:
      type: object
      properties:
        field:
          type: string
        constraint:
          type: string
        message:
          type: string
    PageMeta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasNext:
          type: boolean
      required:
      - total
      - page
      - limit
      - hasNext
...
    CompetitionStatus:
      type: string
      enum:
      - draft
      - published
      - cancelled
    Competition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        coachId:
          type: string
          format: uuid
        name:
          type: string
        discipline:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
        status:
          $ref: '#/components/schemas/CompetitionStatus'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - id
      - coachId
      - name
      - startDate
      - status
    CompetitionCreate:
      type: object
      properties:
        name:
          type: string
        discipline:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
        status:
          $ref: '#/components/schemas/CompetitionStatus'
      required:
      - name
      - startDate
    CompetitionUpdate:
      type: object
      properties:
        name:
          type: string
        discipline:
          type: string
        location:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        description:
          type: string
        status:
          $ref: '#/components/schemas/CompetitionStatus'
    CompetitionEntryStatus:
      type: string
      enum:
      - engaged
      - confirmed
      - withdrawn
    CompetitionEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        competitionId:
          type: string
          format: uuid
        athleteId:
          type: string
          format: uuid
        eventLabel:
          type: string
        status:
          $ref: '#/components/schemas/CompetitionEntryStatus'
        competition:
          $ref: '#/components/schemas/Competition'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
      - id
      - competitionId
      - athleteId
      - status
    EngageRequest:
      type: object
      properties:
        athleteIds:
          type: array
          items:
            type: string
            format: uuid
          minItems: 1
        eventLabel:
          type: string
      required:
      - athleteIds
    CompetitionEntryList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/CompetitionEntry'
      required:
      - data
...
    CompetitionPage:
      type: object
      required:
      - data
      - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Competition'
        meta:
          $ref: '#/components/schemas/PageMeta'
...