# packages/api-client/src/generated/model/competition.ts
/**
 * Generated by orval v8.15.0 🍺
 * Do not edit manually.
 * Talent-X API
 * Contrat d'API REST de Talent-X — source de vérité dérivée de TX-SPEC-002 v2.0 (catalogue d'endpoints, matrice d'autorisation, conventions et format d'erreur) et alignée sur le modèle de données TX-DATA-006.
 *
 * Conventions transverses : préfixe /api/v1 ; jeton d'accès JWT (RS256) via en-tête Authorization ; pagination par enveloppe { data, meta } ; idempotence des écritures sensibles via Idempotency-Key ; opérations longues asynchrones (202 + ressource de statut) ; rate limiting signalé par les en-têtes RateLimit-*. L'autorisation combine rôle, appartenance (lien coach↔athlète), propriété et consentement ; voir TX-SPEC-002 §6.
 * OpenAPI spec version: 1.0.0
 */
import type { CompetitionStatus } from './competitionStatus';

export interface Competition {
  id: string;
  coachId: string;
  name: string;
  discipline?: string;
  location?: string;
  startDate: string;
  endDate?: string;
  description?: string;
  status: CompetitionStatus;
  createdAt?: string;
  updatedAt?: string;
}